feat: 온도 기능 생성

+++ src/main/java/com/example/demo/weather/application/WeatherService.java
@@ -3,6 +3,6 @@ package com.example.demo.weather.application;
 import java.util.List;
 
 public interface WeatherService {
-    List<Integer> tempYear(int year);
-    List<Integer> tempMonth(int year);
+    List<Double> tempYear(int year);
+    List<Double> tempMonth(int year);
 }
